+++ ft_calloc.c
@@ -1,10 +1,10 @@
 #include "libft.h"
 
-void	*ft_calloc(unsigned int count, unsigned int size)
+void	*ft_calloc(size_t count, size_t size)
 {
-	unsigned int total;
-	unsigned char *ptr;
-	unsigned int i;
+	size_t	total;
+	unsigned char	*ptr;
+	size_t	i;
 
 	total = count * size;
 	ptr = (unsigned char *)malloc(total);
