+++ ft_memcpy.c
@@ -1,6 +1,6 @@
 #include "libft.h"
 
-void	*ft_memcpy(void *dest, const void *src, unsigned int n)
+void	*ft_memcpy(void *dest, const void *src, size_t n)
 {
 	unsigned char *a;
 	const unsigned char *b;
