+++ ft_memmove.c
@@ -1,6 +1,6 @@
 #include "libft.h"
 
-void*	ft_memmove(void* dest, const void* src, unsigned int n)
+void*	ft_memmove(void* dest, const void* src, size_t n)
 {
 	unsigned char	*a;
 	unsigned const char	*b;
