+++ ft_strlen.c
@@ -1,8 +1,7 @@
 #include "libft.h"
-
-unsigned int	ft_strlen(const char *str)
+size_t	ft_strlen(const char *str)
 {
-	int	i;
+	size_t	i;
 	
 	i = 0;
 	while (str[i] != '\0')
